+++ src/platform/unsupported/model_runner-platform-device.adb
@@ -136,4 +136,6 @@ package body Model_Runner.Platform.Device is
    function Storage_Limit (Item : Context) return Interfaces.Unsigned_64
    is (Item.Storage);
 
+   function Timestamp_Period (Item : Context) return Float is (Item.Tick);
+
 end Model_Runner.Platform.Device;